Simplified consumers & publishers

+++ src/Cron/MessageQueue/Retry.php
@@ -2,6 +2,7 @@
 
 namespace Synerise\Integration\Cron\MessageQueue;
 
+use Magento\Framework\MessageQueue\MessageEncoder;
 use Psr\Log\LoggerInterface;
 use Magento\Framework\MessageQueue\PublisherInterface;
 use Synerise\Integration\Model\ResourceModel\MessageQueue\Retry\CollectionFactory;
@@ -24,21 +25,20 @@ class Retry
     private $collectionFactory;
 
     /**
-     * @var \Magento\Framework\MessageQueue\MessageEncoder
+     * @var MessageEncoder
      */
     private $messageEncoder;
 
     public function __construct(
         LoggerInterface $logger,
         PublisherInterface $publisher,
-        CollectionFactory $collectionFactory
+        CollectionFactory $collectionFactory,
+        MessageEncoder $messageEncoder
     ){
         $this->logger = $logger;
         $this->publisher = $publisher;
         $this->collectionFactory = $collectionFactory;
-
-        $this->messageEncoder = \Magento\Framework\App\ObjectManager::getInstance()->get('\Magento\Framework\MessageQueue\MessageEncoder');
-
+        $this->messageEncoder = $messageEncoder;
     }
 
     public function execute()
@@ -47,16 +47,18 @@ class Retry
             ->addFieldToFilter('created_at', ['lt' => 'NOW() - INTERVAL 5 MINUTE'])
             ->setPageSize(500);
 
-        foreach ($collection as $item) {
-            try {
-                $this->publisher->publish(
-                    $item->getTopicName(),
-                    $this->messageEncoder->decode($item->getTopicName(), $item->getBody())
-                );
+        if($collection->getSize()) {
+            foreach ($collection as $item) {
+                try {
+                    $this->publisher->publish(
+                        $item->getTopicName(),
+                        $this->messageEncoder->decode($item->getTopicName(), $item->getBody())
+                    );
+                } catch(\Exception $e) {
+                    $this->logger->error($e);
+                }
 
                 $item->delete();
-            } catch(\Exception $e) {
-                $this->logger->error($e);
             }
         }
     }
